% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_names.R
\name{get_full_names}
\alias{get_full_names}
\title{Get a list of fullname which can be matched to usernames}
\usage{
get_full_names()
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get a list of fullname which can be matched to usernames
}
