% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usage.R
\name{get_usage}
\alias{get_usage}
\title{Get the space usage of a given drive}
\usage{
get_usage(drive)
}
\arguments{
\item{drive}{a drive name}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get the space usage of a given drive
}
