% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_usage.R
\name{plot_usage}
\alias{plot_usage}
\title{Plot usage data}
\usage{
plot_usage(usage_data, drive_name)
}
\arguments{
\item{usage_data}{data provided by \code{\link[=get_usage]{get_usage()}}}

\item{drive_name}{a drive name}
}
\value{
a plot object from \link[tibble:tibble-package]{ggplot}
}
\description{
Plot usage data
}
