% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_membership.R
\name{get_group_membership}
\alias{get_group_membership}
\title{Get all groups or users in a given group}
\usage{
get_group_membership(group_name = NULL)
}
\arguments{
\item{group_name}{name of a group, this is usually the same as a drive name}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get all groups or users in a given group
}
