% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_usage.R
\name{get_usage}
\alias{get_usage}
\title{Get the space usage of a given drive}
\usage{
get_usage(drive)
}
\arguments{
\item{drive}{a drive name e.g. "hscdiip" for "/conf/hscdiip"}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Given a drive name e.g. "hsciipd", "social-care",
"linkage/output" etc. it will return the usage data. This will include
all files, their locations, sizes, owners (with full name), and modification
dates.
}
