% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_usage.R
\name{plot_usage}
\alias{plot_usage}
\title{Plot usage data}
\usage{
plot_usage(usage_data, plot_title = NULL, by = c("user", "folder", "both"))
}
\arguments{
\item{usage_data}{data provided by \code{\link[=get_usage]{get_usage()}}}

\item{plot_title}{The title to use for the plot, by default it will try to
guess the drive name and use that.}

\item{by}{one of "user" (default), "folder" or "all" to split the plot by}
}
\value{
a \link[ggplot2:ggplot2-package]{ggplot} plot object
}
\description{
Creates a plot of the data provided by \code{\link[=get_usage]{get_usage()}}.
}
\seealso{
\code{\link[=get_usage]{get_usage()}}
}
